Crate twig

Twig Templating for Rust

Work in progress - This library is still in development and not yet ready for use. Take a look at the CHANGELOG for more details.

Twig Rust is a template engine for everyone writing web applications with Rust. It is derived from Twig (for PHP) and intended to become a fully compatible port - as far as it makes sense.

By design Twig is

flexible

fast

and secure

Getting Started

use twig::{Setup, Engine};

let setup = Setup::default();
let twig = setup.init_engine().unwrap();
// ..

Syntax and Semantics

Twig uses a syntax similar to the Django and Jinja template languages which inspired the Twig runtime environment.

<!DOCTYPE html>
<html>
    <head>
        <title>Display a thread of posts</title>
    </head>
    <body>
    <h1>{{ thread.title }}</h1>
    <ul>
      {% for post in thread.posts %}
        <li>{{ post }}</li>
      {% endfor %}
    </ul>
    {# note: this comment will be ignored #}
    </body>
</html>

Take a look at this introduction: Twig for template designers.

Flexible Architecture

Twig is designed to be highly extensible:

the Twig compiler only defines general semantics and a very flexible extension mechanism.

extensions define specific behavior and data transformations (like if-statement, for-loop, escape-filter, multiplication-operator, call-expression, etc.)

extensions are chosen at runtime.

if you don't like the default behavior (like if-statement, or call-expression) or if you are missing some functionality (like helpers for a new target like excel-files), all you need to do is replace, add or change extensions.

License

Twig-Rust is released under the new BSD license (code and documentation) - as is the original Twig for PHP.

Reexports

pub use engine::Engine;

pub use setup::Setup;

Modules

api
The extensions API for the Twig Template Engine.

engine
The Twig Engine.

extension
Twig Extensions

loader
Twig template loader.

setup
Stores the Twig configuration.

template
Twig Templates.

Macros

loc!
Expands to the location from which it was invoked.

traced_err!
Creates a traced error for the location from which it was invoked.

try_traced!
Equivalent to try! for traced errors.