Struct twig::api::error::Traced

pub struct Traced<T> where T: Error {
    // some fields omitted
}

Generic error with backtrace.

Wrapper around some error type T implementing std::error::Error. * Adds support for a backtrace. * Automatically derefs to the inner error.

Methods

impl<T> Traced<T> where T: Error

fn new(error: T, location: Location) -> Traced<T>

Create a new twig error out of some generic error code.

Examples

use twig::api::error;
use std::env::VarError;

error::Traced::new(VarError::NotPresent, loc!()); // shorthand: `trace!(VarError::NotPresent)`

fn error(&self) -> &T

Return the first location the error occured.

fn location(&self) -> Option<&Location>

Return the first location the error occured.

fn backtrace(&self) -> &[Location]

fn trace<R>(self, loc: Location) -> Traced<R> where T: Into<R>, R: Error

fn iter(&self) -> ErrorIter

Creates an iterator to iterate along the error cause-chain.