Module twig::api::ext

Twig Extension API

Extensions can define new behavior during the compilation process, via token parser and node visitor. They can also define new node types in the abstract syntax tree, like test, unary operator, binary operator, function, filter, global

See twig::extension::core for example implementations. Note: The core extension is not yet fully implemented (see [CHANGELOG][changelog]).

Modules

op

token_parser

Structs

BinaryOperator
Can be used in variable expressions to combine two results.

UnaryOperator
Can be used in variable expressions to process results.

Traits

Extension
Extends the Twig Engine with new behaviour.

Filter
Can modify the result of variable expressions.

Function
Can be used to perform complex computations.

Global
Can be used to define global constants.

NodeVisitor
Modifies the abstract syntax tree immediately after parsing.

Test
Can be used in conditional Twig statements.

TokenParser
Transforms a sub-sequence from the token stream (=lexed template) to nodes in the abstract syntax tree.