Struct twig::engine::extension_registry::ExtensionRegistry

pub struct ExtensionRegistry {
    // some fields omitted
}

Methods

impl ExtensionRegistry

fn new<I>(iterable: I, options: &Options) -> Result<Self, Traced<ExtensionRegistryError>> where I: IntoIterator<Item=Box<Extension>>

Initialize new extension registry instance.

fn has(&self, name: &str) -> bool

Returns true if the given extension is registered

fn ext_names(&self) -> &HashSet<String>

Get all registered extension names.

fn token_parsers(&self) -> &HashMap<String, Box<TokenParser>>

Get the token parser instances defined by engine extensions.

fn node_visitors(&self) -> &Vec<Box<NodeVisitor>>

Get the node visitor instances defined by engine extensions.

fn filters(&self) -> &HashMap<String, Box<Filter>>

Get the filters defined by engine extensions.

fn tests(&self) -> &HashMap<String, Box<Test>>

Get the tests defined by engine extensions.

fn functions(&self) -> &HashMap<String, Box<Function>>

Get the functions defined by engine extensions.

fn operators_unary(&self) -> &HashMap<String, UnaryOperator>

Get the unary operators defined by engine extensions.

fn operators_binary(&self) -> &HashMap<String, BinaryOperator>

Get the binary operators defined by engine extensions.

fn _globals(&self) -> &Vec<Box<Global>>

Get the global variables defined by engine extensions.