Struct twig::engine::parser::lexer::job::cursor::Cursor

pub struct Cursor<'a> {
    // some fields omitted
}

Methods

impl<'a> Cursor<'a>

fn new(template: &'a Raw) -> Cursor<'a>

fn move_by(&mut self, increment: usize)

move the cursor position by increment and keep track of the line

increment in bytes (not chars)

panics

when the increment would move the cursor position out of range

fn move_to(&mut self, pos: usize)

moves the cursor to a new position

Panics

Panics if pos is greater than the template's length.

fn slice_by(&mut self, len: usize) -> &'a str

fn slice_to_end(&mut self) -> &'a str

fn slice_to(&mut self, pos: usize) -> &'a str

fn _template(&self) -> &Raw

fn tail(&self) -> &'a str

fn is_eof(&self) -> bool

fn position(&self) -> usize

fn line(&self) -> usize

fn column(&self) -> usize

fn set_line(&mut self, line: usize)