Module twig::engine::parser::lexer::patterns

The patterns used by the lexer to tokenize the templates.

Reexports

pub use self::options::Options;

Modules

block_end
The block pattern used by the lexer to tokenize the templates.

block_line
The block line pattern used by the lexer to tokenize the templates.

comment_end
The comment pattern used by the lexer to tokenize the templates.

expression_end
The expression pattern used by the lexer to tokenize the templates.

interpolation_end
The interpolation end pattern used by the lexer to tokenize the templates.

interpolation_start
The interpolation start pattern used by the lexer to tokenize the templates.

macros
Rust macros

name
The name pattern used by the lexer to tokenize the templates.

number
The name pattern used by the lexer to tokenize the templates.

operator
The operator pattern used by the lexer to tokenize the templates.

options
Options of the lexer.

punctuation
The name pattern used by the lexer to tokenize the templates.

string
The name pattern used by the lexer to tokenize the templates.

string_dq_delim
The name pattern used by the lexer to tokenize the templates.

string_dq_part
The name pattern used by the lexer to tokenize the templates.

token_start
The tokens_start pattern used by the lexer to tokenize the templates.

verbatim_end
The verbatim_end pattern used by the lexer to tokenize the templates.

verbatim_start
The block raw pattern used by the lexer to tokenize the templates.

Structs

ExtractIter

Patterns

Traits

Extract

Functions

_php_trim_right
Trim whitespace in a php-compatible manner

php_stripcslashes