Struct twig::engine::setup::Setup

pub struct Setup {
    // some fields omitted
}

Methods

impl Setup

Builds an instance of the Twig Engine, according to supplied options and engine extensions.

The following extensions will be registered by default: * core * escaper * optimizer

Examples

use twig::{Setup, Engine};
use twig::extension::Debug;

let mut setup = Setup::default()
    .set_strict_variables(true)
    .add_extension(Debug::new()).unwrap();
let engine = Engine::new(setup).unwrap();

fn engine(self) -> Result<Engine, Traced<TwigError>>

Create engine from setup.

Examples

use twig::Setup;

let twig = Setup::default().engine().unwrap();

fn add_extension(self, extension: Box<Extension>) -> Result<Self, Traced<TwigError>>

Registers an extension

fn set_debug(self, debug: bool) -> Self

When set to true, it automatically set "auto_reload" to true as well (default to false)

fn set_charset(self, set_charset: Charset) -> Self

The charset used by the templates (default to UTF-8)

fn set_strict_variables(self, strict_variables: bool) -> Self

Whether to ignore invalid variables in templates (default to false).

fn set_autoescape(self, autoescape: Autoescape) -> Self

Whether to enable auto-escaping (default to html): * false: disable auto-escaping * true: equivalent to html * html, js: set the autoescaping to one of the supported strategies * filename: set the autoescaping strategy based on the template filename extension * PHP callback: a PHP callback that returns an escaping strategy based on the template "filename"

fn set_cache(self, cache: Option<&Path>) -> Self

An absolute path where to store the compiled templates (optional)

fn set_auto_reload(self, auto_reload: Option<bool>) -> Self

Whether to reload the template if the original source changed (optional). If you don't provide the auto_reload option, it will be determined automatically based on the debug value.

fn set_optimizations(self, optimizations: Optimizations) -> Self

A flag that indicates whether optimizations are applied

fn options(&self) -> &Options

Get all options

fn extensions(&self) -> Iter

Get all registered extensions