Struct twig::instructions::CompiledExpression

pub struct CompiledExpression {
    // some fields omitted
}

Represents a mess created by a compiled expression.

It is up to the caller to clean up this mess by calling finalize on this struct.

Methods

impl CompiledExpression

fn with_result(origin: &'static str, result: Mem) -> CompiledExpression

fn empty(origin: &'static str) -> CompiledExpression

fn new(origin: &'static str, result: Mem, stack_length: u16) -> CompiledExpression

fn result(&self) -> Option<Mem>

fn finalize<'c, 'r>(self, stage: &'r mut Staging<'c, Value>) -> TemplateResult<()>