Trait twig::loader::Loader

pub trait Loader: Debug {
    fn source<'a>(&'a mut self, name: &str) -> Result<Cow<str>, Traced<LoaderError>>;
    fn cache_key<'a>(&'a mut self, name: &str) -> Result<Cow<'a, str>, Traced<LoaderError>>;
    fn is_fresh(&mut self, name: &str, time: i64) -> bool;
}

Required Methods

fn source<'a>(&'a mut self, name: &str) -> Result<Cow<str>, Traced<LoaderError>>

Gets the source code of a template, given its name

Returns a Cow to allow for efficient caching mechanisms.

Failures

When name is not found

fn cache_key<'a>(&'a mut self, name: &str) -> Result<Cow<'a, str>, Traced<LoaderError>>

Gets the cache key to use for the cache for a given template

Failures

When name is not found

fn is_fresh(&mut self, name: &str, time: i64) -> bool

returns true if the template is still fresh