Module twig::nodes

Produces an abstract syntax tree from a stream of tokens.

Modules

body

body_parser

expr

expr_parser

module_parser

Structs

Block

EmbededTemplate

ImportedFunction

Macro

Module

Parser
Helpers for manipulating and inspecting token iterators when creating AST.

TokenParser

Trait

Traits

Parse

TokenParserExtension

Functions

parse
Parse given token stream into a node tree.