Trait twig::template::api::Template

pub trait Template {
    fn render(&self, runtime: &Runtime) -> Result<String, Traced<TwigError>>;
    fn display(&self, runtime: &Runtime, blocks: Option<Vec<()>>);
}

Required Methods

fn render(&self, runtime: &Runtime) -> Result<String, Traced<TwigError>>

Renders the template with the given context and returns it as string.

fn display(&self, runtime: &Runtime, blocks: Option<Vec<()>>)

Displays the template with the given context.

context is an array of parameters to pass to the template blocks is an array of blocks to pass to the template