Crate twig

Twig library for rust.

Examples

use twig::engine::Setup;
use twig::loader;
use twig::runtime::Runtime;
use twig::template::api::Template;

let mut loader = loader::array::Array::default();
loader.set_template("greetings","Hello {{name}}!");

let mut engine = Setup::default().engine().unwrap();
engine.set_loader(Box::new(loader));

let mut runtime = Runtime::default();
runtime.set("name", "world");

let compiled = engine.load_template("greetings", None).unwrap();
assert_eq!(&compiled.render(&runtime).unwrap(), "Hello world!")

Reexports

pub use engine::Engine;

pub use engine::Setup;

Modules

api
The extensions API for the Twig Template Engine.

engine
The Twig Engine

extension
Twig Extensions

loader
Twig template loader.

runtime
Runtime for compiled templates

template
Twig Templates.

Macros

loc!
Expands to the location from which it was invoked.

traced_err!
Creates a traced error for the location from which it was invoked.

try_traced!
Equivalent to try! for traced errors.

try_untraced!